Finished moving everything, now cleaning up docs
