--- src/traits.rs
@@ -1,1 +1,0 @@
-
